usa 32bit result for matrix C

+++ include/picongpu/fields/currentDeposition/Esirkepov/Esirkepov.hpp
@@ -164,8 +164,8 @@ struct Esirkepov<T_ParticleShape, DIM3>
 #endif
         nvcuda::wmma::fragment<nvcuda::wmma::matrix_a, 16, 16, 16, half, nvcuda::wmma::col_major> a_frag;
         nvcuda::wmma::fragment<nvcuda::wmma::matrix_b, 16, 16, 16, half, nvcuda::wmma::row_major> b_frag;
-        nvcuda::wmma::fragment<nvcuda::wmma::accumulator, 16, 16, 16, half> acc_frag;
-        nvcuda::wmma::fill_fragment(acc_frag, __float2half(0.0));
+        nvcuda::wmma::fragment<nvcuda::wmma::accumulator, 16, 16, 16, float> acc_frag;
+        nvcuda::wmma::fill_fragment(acc_frag, float(0.0));
         nvcuda::wmma::load_matrix_sync(a_frag, matA.getPointer(), 16);
         nvcuda::wmma::load_matrix_sync(b_frag, matB.getPointer(), 16);
         nvcuda::wmma::mma_sync(acc_frag, a_frag, b_frag, acc_frag);
@@ -382,7 +382,7 @@ struct Esirkepov<T_ParticleShape, DIM3>
                     );
                     float_X accumulated_J( 0.0 );
 
-                    float_X const tmp = -currentSurfaceDensity[ w ] * __half2float( matResult( id2d + offset ) );
+                    float_X const tmp = -currentSurfaceDensity[ w ] * ( matResult( id2d + offset ) );
                     /* we cheat a little bit y and z can be set both to the id2d.x
                      * because the correct direction will be later overwritten with k
                      */
@@ -403,7 +403,7 @@ struct Esirkepov<T_ParticleShape, DIM3>
                 {
                     float_X accumulated_J( 0.0 );
                     DataSpace< DIM2 > matOffset = DataSpace< DIM2 >( b * 8 , 4 + b * 8 );
-                    float_X const tmp = -currentSurfaceDensity.x() * __half2float( matResult( id2d + matOffset ) );
+                    float_X const tmp = -currentSurfaceDensity.x() * ( matResult( id2d + matOffset ) );
                     DataSpace< DIM3 > jOffset( 0, begin + id2d.y(), begin + id2d.x() );
                     for( int k = begin ; k < end; ++k )
                     {
